browse docs inside connect side page
